Refactor: extract exam error handling to dedicated service
- Created ExamErrorHandler service with static methods
- Moved getExamErrorMessage() and extractErrorKey() from navigationCommands
- Improved code organization and reusability
- Fixed linting issues (added curly braces for control statements)
- Reduced navigationCommands.ts from 561 to 446 lines
